+++ src/main.rs
@@ -6,26 +6,21 @@
 //! with support for resumable downloads and MD5 hash verification.
 
 use ia_get::{IaGetError, Result};
-use ia_get::utils::{create_progress_bar, create_spinner, format_duration, format_size, format_transfer_rate};
+use ia_get::utils::create_spinner; // Removed format_duration, format_size, format_transfer_rate
+use ia_get::downloader; 
 use indicatif::ProgressStyle;
 use regex::Regex;
-use reqwest::header::{HeaderValue, HeaderMap};
+// Removed use reqwest::header::{HeaderValue, HeaderMap};
 use reqwest::Client;
 use serde_xml_rs::from_str;
 use clap::Parser;
-use std::fs::{self, File};
-use std::io::{BufReader, Read, Write};
+// Removed use std::fs::{self, File};
+// Removed use std::io::{BufReader, Read, Write};
 use std::process;
-use std::path::Path;
+// Removed use std::path::Path;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use ia_get::archive_metadata::XmlFiles;
-
-/// Buffer size for file operations (8KB)
-const BUFFER_SIZE: usize = 8192;
-
-/// File size threshold for showing hash progress bar (16MB)
-const LARGE_FILE_THRESHOLD: u64 = 16 * 1024 * 1024;
 
 /// User agent string for HTTP requests
 const USER_AGENT: &str = "ia-get";
@@ -78,329 +73,6 @@ fn get_xml_url(original_url: &str) -> String {
     } else {
         base_new_url
     }
-}
-
-/// Calculates the MD5 hash of a file
-/// 
-/// Uses a streaming approach to compute the MD5 hash by reading the file in chunks,
-/// which is memory-efficient for large files. This approach avoids loading the entire
-/// file into memory at once.
-/// 
-/// # Arguments
-/// * `file_path` - Path to the file to hash
-/// * `running` - Signal handler to check for interruption
-/// 
-/// # Returns
-/// * `Ok(String)` - The MD5 hash as a lowercase hexadecimal string
-/// * `Err(IaGetError)` - If the file cannot be read
-fn calculate_md5(file_path: &str, running: &Arc<AtomicBool>) -> Result<String> {
-    let file = File::open(file_path)?;
-    let file_size = file.metadata()?.len();
-    let is_large_file = file_size > LARGE_FILE_THRESHOLD;
-    
-    let mut reader = BufReader::with_capacity(BUFFER_SIZE, file);
-    let mut context = md5::Context::new();
-    let mut buffer = [0; BUFFER_SIZE];
-    
-    // Only show progress bar for large files to avoid UI clutter
-    let pb = if is_large_file {
-        let progress_bar = create_progress_bar(file_size, "╰╼ Hashing      ", Some("cyan/blue"), false);
-        Some(progress_bar)
-    } else {
-        None
-    };
-    
-    let mut bytes_processed: u64 = 0;
-    
-    loop {
-        let bytes_read = reader.read(&mut buffer)?;
-        if bytes_read == 0 {
-            // End of file reached
-            break;
-        }
-        
-        // Check for signal interruption during hash calculation
-        if !running.load(Ordering::SeqCst) {
-            if let Some(ref progress_bar) = pb {
-                progress_bar.finish_and_clear();
-            }
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::Interrupted,
-                "Hash calculation interrupted by signal",
-            ).into());
-        }
-        
-        context.consume(&buffer[..bytes_read]);
-        
-        // Update progress if we're showing it
-        if let Some(ref progress_bar) = pb {
-            bytes_processed += bytes_read as u64;
-            progress_bar.set_position(bytes_processed);
-        }
-    }
-    
-    // Finalize progress bar if it exists
-    if let Some(progress_bar) = pb {
-        progress_bar.finish_and_clear();
-    }
-    
-    let hash = context.compute();
-    Ok(format!("{:x}", hash))
-}
-
-/// Check if an existing file has the correct hash
-///
-/// # Arguments
-/// * `file_path` - Path to the file to check
-/// * `expected_md5` - Expected MD5 hash, if available
-/// * `running` - Signal handler to check for interruption
-///
-/// # Returns
-/// * `Ok(Some(bool))` - Whether the file exists and its hash matches or not
-/// * `Ok(None)` - If the file doesn't exist
-/// * `Err(IaGetError)` - If there was an error checking the file
-fn check_existing_file(file_path: &str, expected_md5: Option<&str>, running: &Arc<AtomicBool>) -> Result<Option<bool>> {
-    if !Path::new(file_path).exists() {
-        return Ok(None);
-    }
-
-    if expected_md5.is_none() {
-        return Ok(Some(true)); // No MD5 to check against, assume file is valid
-    }
-
-    // Calculate the MD5 hash of the local file
-    let local_md5 = match calculate_md5(file_path, running) {
-        Ok(hash) => hash,
-        Err(e) => {
-            // Check if this is an interruption by looking at the error message
-            if e.to_string().contains("interrupted by signal") {
-                return Err(e);
-            }
-            println!("╰╼ Failed to calculate MD5 hash: {}", e);
-            return Ok(Some(false));
-        }
-    };
-
-    Ok(Some(local_md5 == expected_md5.unwrap()))
-}
-
-/// Ensure parent directories exist for a file
-///
-/// # Arguments
-/// * `file_path` - Path to the file
-///
-/// # Returns
-/// * `Ok(())` - If directories were created or already exist
-/// * `Err(IaGetError)` - If directories couldn't be created
-fn ensure_parent_directories(file_path: &str) -> Result<()> {
-    // Check if file_name includes a path
-    if let Some(path) = Path::new(file_path).parent() {
-        // Create the local directory if it doesn't exist and path has a file name
-        if path.file_name().is_some() && !path.exists() {
-            fs::create_dir_all(path)?;
-        }
-    }
-    Ok(())
-}
-
-/// Prepare a file for download, potentially resuming an existing download
-///
-/// # Arguments
-/// * `file_path` - Path to the file
-///
-/// # Returns
-/// * `Ok(File)` - Open file handle ready for writing
-/// * `Err(IaGetError)` - If the file couldn't be opened
-fn prepare_file_for_download(file_path: &str) -> Result<File> {
-    let file = std::fs::OpenOptions::new()
-        .write(true)
-        .create(true)
-        .truncate(false)
-        .open(file_path)?;
-    
-    Ok(file)
-}
-
-/// Download file content with progress reporting
-///
-/// # Arguments
-/// * `client` - HTTP client
-/// * `url` - URL to download from
-/// * `file_size` - Current size of local file (for resuming)
-/// * `file` - Open file handle to write to
-/// * `running` - Signal handler to check for interruption
-/// * `is_resuming` - Whether this is a resumed download or fresh download
-///
-/// # Returns
-/// * `Ok(u64)` - Total bytes downloaded
-/// * `Err(IaGetError)` - If download failed
-async fn download_file_content(
-    client: &Client, 
-    url: &str, 
-    file_size: u64, 
-    file: &mut File,
-    running: &Arc<AtomicBool>,
-    is_resuming: bool
-) -> Result<u64> {
-    let download_action = if is_resuming { "╰╼ Resuming     " } else { "╰╼ Downloading  " };
-    
-    // Set the Range header to specify the starting offset
-    let mut initial_request = client.get(url);
-    let range_header = format!("bytes={}-", file_size);
-    let mut headers = HeaderMap::new();
-    // Fixed: Handle the HeaderValue::from_str error explicitly
-    headers.insert(
-        reqwest::header::RANGE, 
-        HeaderValue::from_str(&range_header).map_err(|e| IaGetError::UrlError(format!("Invalid header value: {}", e)))?
-    );
-    initial_request = initial_request.headers(headers);
-
-    let mut response = initial_request.send().await?;
-
-    // Get the content length from the response headers
-    let content_length = response.content_length().unwrap_or(0);
-    let pb = create_progress_bar(
-        content_length + file_size,
-        download_action,
-        None, // Default to green/green
-        true  // Include ETA
-    );
-
-    // Record start time to calculate transfer rate later
-    let start_time = std::time::Instant::now();
-
-    // Download the remaining chunks and update the progress bar
-    let mut total_bytes: u64 = file_size;
-    let mut downloaded_bytes: u64 = 0;
-    
-    while let Some(chunk) = response.chunk().await? {
-        // Check for signal interruption during download
-        if !running.load(Ordering::SeqCst) {
-            pb.finish_and_clear();
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::Interrupted,
-                "Download interrupted during file transfer",
-            ).into());
-        }
-        
-        file.write_all(&chunk)?;
-        downloaded_bytes += chunk.len() as u64;
-        total_bytes += chunk.len() as u64;
-        pb.set_position(total_bytes);
-    }
-
-    // Calculate elapsed time and transfer rate
-    let elapsed = start_time.elapsed();
-    let elapsed_secs = elapsed.as_secs_f64();
-    
-    // Calculate transfer rate (bytes per second)
-    let transfer_rate = if elapsed_secs > 0.0 {
-        downloaded_bytes as f64 / elapsed_secs
-    } else {
-        0.0 // Avoid division by zero
-    };
-    
-    // Format the transfer rate
-    let (rate, unit) = format_transfer_rate(transfer_rate);
-    
-    // Clear the progress bar and show completion details
-    pb.finish_and_clear();
-    
-    // Show completion message with size, time and speed
-    println!("├╼ Downloaded   ⤵️ {} in {} ({:.2} {}/s)", 
-        format_size(downloaded_bytes),
-        format_duration(elapsed),
-        rate,
-        unit);
-
-    
-    Ok(total_bytes)
-}
-
-/// Verify a downloaded file's hash against an expected value
-///
-/// # Arguments
-/// * `file_path` - Path to the file
-/// * `expected_md5` - Expected MD5 hash, if available
-/// * `running` - Signal handler to check for interruption
-///
-/// # Returns
-/// * `Ok(bool)` - Whether the hash matches
-/// * `Err(IaGetError)` - If verification failed
-fn verify_downloaded_file(file_path: &str, expected_md5: Option<&str>, running: &Arc<AtomicBool>) -> Result<bool> {   
-    // Calculate the MD5 hash of the local file
-    let local_md5 = match calculate_md5(file_path, running) {
-        Ok(hash) => hash,
-        Err(e) => {
-            println!("╰╼ Failed to calculate MD5 hash: {}", e);
-            return Ok(false);
-        }
-    };
-    
-    match expected_md5 {
-        Some(expected) => {
-            let matches = local_md5 == expected;
-            if !matches {
-                println!("╰╼ Hash         ❌");
-            } else {
-                println!("╰╼ Hash         ✅");
-            }
-            Ok(matches)
-        },
-        None => {
-            println!("╰╼ No MD5:      ⚠️");
-            Ok(true) // Consider it valid if no expected hash
-        },
-    }
-}
-
-/// Download a file from archive.org with resume capability
-///
-/// # Arguments
-/// * `client` - HTTP client
-/// * `url` - URL to download from
-/// * `file_path` - Path to save the file
-/// * `expected_md5` - Expected MD5 hash, if available
-/// * `running` - Signal handler to check for interruption
-///
-/// # Returns
-/// * `Ok(())` - If download was successful or file already exists with correct hash
-/// * `Err(IaGetError)` - If download failed
-async fn download_file(
-    client: &Client,
-    url: &str,
-    file_path: &str,
-    expected_md5: Option<&str>,
-    running: &Arc<AtomicBool>
-) -> Result<()> {
-    println!(" ");
-    println!("📦️ Filename     {}", file_path);
-    
-    // Check if the file exists and has correct hash
-    if let Some(is_valid) = check_existing_file(file_path, expected_md5, running)? {
-        if is_valid {
-            println!("╰╼ Completed:   ✅");
-            return Ok(());
-        }
-    }
-    
-    // Create directories if needed
-    ensure_parent_directories(file_path)?;
-    
-    // Open the file for writing/appending
-    let mut file = prepare_file_for_download(file_path)?;
-    
-    // Get the size of the local file if it already exists
-    let file_size = file.metadata()?.len();
-    let is_resuming = file_size > 0;
-    
-    // Download the file content
-    download_file_content(client, url, file_size, &mut file, running, is_resuming).await?;
-    
-    // Verify the downloaded file
-    verify_downloaded_file(file_path, expected_md5, running)?;
-    
-    Ok(())
 }
 
 /// Command-line interface for ia-get
@@ -545,7 +217,7 @@ async fn main() -> std::result::Result<(), Box<dyn std::error::Error>> {
         }
 
         // Download the file
-        download_file(
+        downloader::download_file( // Updated to use downloader::download_file
             &client, 
             absolute_url.as_str(), 
             &file.name,
